Starting an actual repo for this.
Some of it is a bit ugly. Has export functionality for tex and md,
but I need to do some stripping on the inputs to get it right.

I also want to look into using json instead of pickling.
Not sure whether the current object structure will translate to
and from json reliably.